% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funjacc.R
\name{funjacc}
\alias{funjacc}
\title{Run FunJacc}
\usage{
funjacc(
  gene_list,
  p_cut = 0.01,
  jaccard_cut = 0.5,
  data_types = c("GO:BP"),
  inflation = 2,
  species = "hsapiens",
  default_node_label_size = 10,
  cluster_label_size = 20
)
}
\arguments{
\item{gene_list}{List of genes in which to look for enriched terms}

\item{p_cut}{P-value cutoff for gProfiler results}

\item{jaccard_cut}{Jaccard index cutoff for identifying high scoring links between terms}

\item{data_types}{List of functional terms to include e.g.c('GO:BP', 'GO:MF', 'GO:CC', 'KEGG', 'REAC', 'TF', 'MIRNA', 'CORUM', 'HP', 'HPA', 'WP'). Supplying 'all' will include all these terms}

\item{inflation}{Inflation parameter for MCL clustering ~0.5-2}

\item{species}{Organism to use for gProfiler to interpret gene list e.g. 'hsapiens', 'mmusculus'}

\item{default_node_label_size}{Node size to use for Cytoscape plotting}

\item{cluster_label_size}{Node size for cluster labels in Cytoscape plotting}
}
\value{
list of results elements 'annotation' = clusters and their annotation, 'network' = network of clusters
}
\description{
Determine enriched functional term clusters with FunJacc
}
\examples{
funjacc(gene_list, data_types=data_types, species='hsapiens', inflation=3)
}
